combined common frontend functions into utils.py
