+++ frontend/.env_example
@@ -1,1 +1,2 @@
 AWS_API_LINK =
+OPENAI_API_KEY =
